Adding removeBooks.js module
